Better Offline mode and Bug Fixes
Offline mode is now finally finished, you can browse your entire library when in offline mode it it's downloaded! And a majority of bugs have been fixed.

Even a severe one that occured with each launch, It was caused by the Cod.SysUtils external library which had a function to get the user name, and It was not well inplemented.